backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==9.0.0
Werkzeug==3.0.3
python-dotenv==1.0.0
requests==2.31.0


